requirements.txt
streamlit
google-cloud-aiplatform
vertexai
google-auth
